issue with dedicatedconnect being called repeatidly from join level resolved using if checks

+++ levelhost.cpp
@@ -25,7 +25,7 @@ void LevelHost::draw(){
             locked->draw(g->getscreenwidth(),g->getscreenheight());
         }else{
             std::vector<std::string> buffers = {"Connecting to matchmaking server"};
-            for(int i = 0; i < buffers.size();++i){
+            for(size_t i = 0; i < buffers.size();++i){
                 Color text = { 130, 130, 130, 255 };
                 if(index == i){
                     text.r = 230;
